+++ tests/state_machine_test.cc
@@ -1,7 +1,6 @@
 ﻿#include "gtest/gtest.h"
 
-extern "C"
-void state_machine_utest(void);
+extern "C" void state_machine_utest(void);
 
 TEST(StateMachine, utest) {
   state_machine_utest();